sphinxsort decomposition part1
